# Performance enhancements