- Fix README NuGet package link 
- Performance enhancements
  - Use pattern match, use array instead of list when possible

Benchmark result, compare with the latest nuget package(2.0.0)

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.22616
Intel Core i5-6300U CPU 2.40GHz (Skylake), 1 CPU, 4 logical and 2 physical cores
.NET Core SDK=7.0.100-preview.4.22252.9
  [Host]     : .NET Core 3.1.18 (CoreCLR 4.700.21.35901, CoreFX 4.700.21.36305), X64 RyuJIT
  DefaultJob : .NET Core 3.1.18 (CoreCLR 4.700.21.35901, CoreFX 4.700.21.36305), X64 RyuJIT


```
|       Method |          Mean |       Error |        StdDev |        Median | Ratio | RatioSD |      Gen 0 | Gen 1 | Gen 2 |   Allocated |
|------------- |--------------:|------------:|--------------:|--------------:|------:|--------:|-----------:|------:|------:|------------:|
| **EnumerateNew** | **26,443.776 μs** | **415.4109 μs** |   **368.2510 μs** | **26,455.742 μs** |  **0.92** |    **0.08** | **22187.5000** |     **-** |     **-** | **34021.69 KB** |
| **EnumerateOld** | **29,629.458 μs** | **896.6912 μs** | **2,469.7480 μs** | **29,110.257 μs** |  **1.00** |    **0.00** | **22142.8571** |     **-** |     **-** | **34021.72 KB** |
|              |               |             |               |               |       |         |            |       |       |             |
|    **Count0New** |      **2.422 μs** |   **0.0933 μs** |     **0.2720 μs** |      **2.319 μs** |  **1.20** |    **0.21** |     **0.9689** |     **-** |     **-** |     **1.48 KB** |
|    **Count0Old** |      **2.055 μs** |   **0.1102 μs** |     **0.3127 μs** |      **2.011 μs** |  **1.00** |    **0.00** |     **0.9880** |     **-** |     **-** |     **1.52 KB** |
|              |               |             |               |               |       |         |            |       |       |             |
|    **Count1New** |      **4.883 μs** |   **0.0945 μs** |     **0.1161 μs** |      **4.898 μs** |  **0.76** |    **0.05** |     **2.7771** |     **-** |     **-** |     **4.26 KB** |
|    **Count1Old** |      **5.785 μs** |   **0.2194 μs** |     **0.6434 μs** |      **5.559 μs** |  **1.00** |    **0.00** |     **2.8000** |     **-** |     **-** |     **4.29 KB** |
|              |               |             |               |               |       |         |            |       |       |             |
|    **Count2New** |      **6.007 μs** |   **0.1149 μs** |     **0.2042 μs** |      **6.034 μs** |  **0.97** |    **0.08** |     **3.1967** |     **-** |     **-** |      **4.9 KB** |
|    **Count2Old** |      **6.241 μs** |   **0.1338 μs** |     **0.3753 μs** |      **6.264 μs** |  **1.00** |    **0.00** |     **3.2120** |     **-** |     **-** |     **4.93 KB** |